fix(webhook): map original images to containers based on container name
When using a tool like linkerd that uses a webhook to add containers, if linkerd webhook is executed after our pod webhook, indexes can change and therefore the mapping becomes incorrect. This commit fix it by mapping on names instead of indexes, but be aware that a webhook that would change the name of containers will cause mapping to fail again.