fix(helm): rename registry statefulset and related resources
It was originally named "cache-registry" at the time the project was also called "cache-registry". Thus, when reworking the chart, its name was replaced by a template to name it by the name of the project. Then, when the project was renamed kube-image-keeper, the registry name was templated as kube-image-keeper instead of cache-registry. This commit intend to fix this mistake and set a proper name for the registry.
cjob