fix(controller): prevent replacing ":" with "-" in CachedImage spec
When a registry url contains a port (eg. myreg.com:5000), the annotation
original image is first populated with the real original image. Then if
the pod is updated, we update this annotation with the current image but
we remove the "localhost:port/" part. So from the original image
myreg.com:5000/my-image we get the rerwitten image
localhost:7439/myreg.com-5000/my-image and the annotation is updated as
myreg.com-5000/my-image, which no longer is equal to the original image.

By checking that the image does not starts with localhost:7439/, we
ensure that the image is the original one, and thus we set the
annotation.