fix: remove dependency to k8s.io/kubernetes/pkg/credentialprovider
This dependency forced us to add k8s.io/kubernetes as a dependency,
which is ill advised. It is not re-published as standalone package like
other k8s packages, so we cannot import it like we do for client-go for
instance. Thus, we copied parts of this package that we use here in
order to remove the dependency.